feat: create SectionMangerView and import to MenuView
